performWithMaxBackoff
Expose a way for callers to specify the maximum backoff exponent when
truncating the exponential backoff delay. This was motivated by the
Pinterest API, which enforces a 1 hour waiting period when a rate-limit
is exceeded. It may not be an ideal solution in many cases, as there's
still a thread of execution being held up while we wait to make another
request, but it does at least prevent numerous pointless requests which
the third-party could perceive as spamming.